move gateway controller to its own folder
