+++ internal/controllers/v1beta1/gateway/gateway_test.go
@@ -1,4 +1,4 @@
-package v1beta1
+package gateway
 
 import (
 	"context"
