+++ internal/controllers/v1beta1/gateway/options.go
@@ -1,4 +1,4 @@
-package v1beta1
+package gateway
 
 type OptionsFunc func(*GatewayController)
